Moved packetqueue to core/net/rime, as it is only used by the Rime collect protocol
